add validation in discipline

+++ src/components/teacherItem/TeacherItem.tsx
@@ -4,13 +4,16 @@ import { TeacherType } from "../../pages/teachers/Teachers";
 import deleteIcon from "../../images/deleteico.svg";
 import updateIcon from "../../images/updateico.svg";
 import closeIcon from "../../images/closeico.svg";
+import showMoreIcon from "../../images/showmoreico.svg";
 import { Modal } from "../modal/Modal";
 import { CustomForm } from "../form/CustomForm";
+import { NavLink } from "react-router-dom";
 
 type TeacherItemPropsType = {
 	teacher: TeacherType;
-	deleteTeacherCB: (teacherId: number) => void;
-	editTeacherCB: (teacherId: number, updatedData: TeacherType) => void;
+	deleteTeacherCB?: (teacherId: number) => void;
+	editTeacherCB?: (teacherId: number, updatedData: TeacherType) => void;
+	full?: boolean;
 };
 
 export const TeacherItem: React.FC<TeacherItemPropsType> = (props: TeacherItemPropsType) => {
@@ -21,7 +24,7 @@ export const TeacherItem: React.FC<TeacherItemPropsType> = (props: TeacherItemPr
 	};
 
 	const onDelete = () => {
-		props.deleteTeacherCB(props.teacher.id);
+		props.deleteTeacherCB && props.deleteTeacherCB(props.teacher.id);
 	};
 
 	const onShow = () => {
@@ -30,10 +33,14 @@ export const TeacherItem: React.FC<TeacherItemPropsType> = (props: TeacherItemPr
 
 	return (
 		<div className={classes.teacherItem}>
-			<Modal show={show} setShow={setShow}>
-				<CustomForm teacher={props.teacher} editTeachersFromCB={props.editTeacherCB} onShow={onShow} />
-			</Modal>
-			{!popup && <div className={classes.invisible} onClick={onPopup}></div>}
+			{props.full && (
+				<>
+					<Modal show={show} setShow={setShow}>
+						<CustomForm teacher={props.teacher} editTeachersFromCB={props.editTeacherCB} onShow={onShow} />
+					</Modal>
+					{!popup && <div className={classes.invisible} onClick={onPopup}></div>}
+				</>
+			)}
 			<div className={classes.keys}>
 				<span>Имя</span>
 				<span>Фамилия</span>
@@ -44,18 +51,25 @@ export const TeacherItem: React.FC<TeacherItemPropsType> = (props: TeacherItemPr
 				<span>{props.teacher.surname}</span>
 				<span>{props.teacher.patronimyc}</span>
 			</div>
-			{popup && (
-				<div className={classes.alternative}>
-					<button onClick={onDelete}>
-						<img src={deleteIcon} alt="delete" />
-					</button>
-					<button onClick={onShow}>
-						<img src={updateIcon} alt="update" />
-					</button>
-					<button onClick={onPopup}>
-						<img src={closeIcon} alt="close" />
-					</button>
-				</div>
+			{props.full && (
+				<>
+					{popup && (
+						<div className={classes.alternative}>
+							<button onClick={onDelete}>
+								<img src={deleteIcon} alt="delete" />
+							</button>
+							<button onClick={onShow}>
+								<img src={updateIcon} alt="update" />
+							</button>
+							<button onClick={onPopup}>
+								<img src={closeIcon} alt="close" />
+							</button>
+							<NavLink to={`/discipline/${props.teacher.id}`}>
+								<img src={showMoreIcon} alt="show more" />
+							</NavLink>
+						</div>
+					)}
+				</>
 			)}
 		</div>
 	);
